Added I2C read and write byte functions

+++ Trik_Devices/trik_softi2c.h
@@ -59,8 +59,8 @@ void I2C_highSCL(uint8_t I2C_NUMBER);
 void I2C_init(uint8_t I2C_NUMBER);
 void I2C_start(uint8_t I2C_NUMBER);
 void I2C_stop(uint8_t I2C_NUMBER);
-uint8_t I2C_read(uint8_t I2C_NUMBER, uint16_t i2c_ack);
-uint16_t I2C_write(uint8_t I2C_NUMBER, uint8_t i2c_dta);
+uint8_t I2C_read(uint8_t I2C_NUMBER, uint8_t i2c_ack);
+uint8_t I2C_write(uint8_t I2C_NUMBER, uint8_t i2c_dta);
 uint8_t I2C_readchar(uint8_t i2c_dev, uint8_t i2c_reg);
 void I2C_writechar(uint8_t i2c_dev, uint8_t i2c_reg, uint8_t i2c_val);
